macOS

+++ libfaad/common.h
@@ -227,10 +227,11 @@ typedef signed long long int64_t;
 typedef unsigned long uint64_t;
 typedef signed long int64_t;
 #endif
-typedef unsigned long uint32_t;
+#include <string.h>
+typedef unsigned int uint32_t;
 typedef unsigned short uint16_t;
 typedef unsigned char uint8_t;
-typedef signed long int32_t;
+typedef signed int int32_t;
 typedef signed short int16_t;
 typedef signed char int8_t;
 # endif
